Flattern TooltipUtil into AttributeUtil
